AI-Backend/requirements.txt
fastapi
uvicorn
torch
torchaudio
numpy
soundfile
pyaudio
webrtcvad
faster-whisper
llama-cpp-python
TTS